v0.3.0

+++ backend/marker/src/comparators/regex_comparator.rs
@@ -1,8 +1,8 @@
-//! A comparator that uses regular expressions to find matches and award marks proportionally.
+//! A comparator that uses regular expressions to find matches and award marks proportionally, where **line order matters**.
 //!
 //! The `RegexComparator` is a flexible tool that allows for pattern matching using regex.
 //! It calculates the ratio of matches found in the student's output against the memo's output
-//! and awards marks based on this percentage.
+//! and awards marks based on this percentage. **Lines are compared in order; only lines at the same position are considered a match.**
 
 use crate::traits::comparator::OutputComparator;
 use crate::types::{TaskResult, Subsection};
@@ -13,6 +13,8 @@ use regex::Regex;
 /// This comparator is ideal for tasks where the correctness of the output can be verified with
 /// a regular expression. It provides a powerful way to validate complex patterns. Marks are awarded
 /// based on the ratio of matches in the student's output compared to the memo's output. **Extra lines in the student output are penalized: the score is multiplied by the ratio of memo lines to student lines if student_lines > memo_lines.**
+///
+/// **Note:** Line order matters. Only lines at the same index in both memo and student outputs are considered for matching.
 pub struct RegexComparator;
 
 impl OutputComparator for RegexComparator {
@@ -76,12 +78,12 @@ impl OutputComparator for RegexComparator {
             }
         } else {
             let ratio = awarded_marks as f32 / total_patterns as f32;
-            (section.value as f32 * ratio).round() as u32
+            (section.value as f32 * ratio).round() as i64
         };
 
         if student_lines.len() > memo_lines.len() && student_lines.len() > 0 {
             let penalty = memo_lines.len() as f32 / student_lines.len() as f32;
-            awarded = (awarded as f32 * penalty).round() as u32;
+            awarded = (awarded as f32 * penalty).round() as i64;
         }
 
         TaskResult {
@@ -104,7 +106,7 @@ mod tests {
         lines.iter().map(|s| s.to_string()).collect()
     }
 
-    fn mock_subsection(value: u32) -> Subsection {
+    fn mock_subsection(value: i64) -> Subsection {
         Subsection {
             name: "Mock Subsection".to_string(),
             value,
